Bugfixes for virtual class mapping.
If the owner of a symbol was a virtual class, replace the owner
with the worker trait symbol.

Prettyprints for some trees still seem to show types containing
old virtual class (e. g. Template subtree of a virtual trait).
I suspect that the types are correctly updated but the name fields
of some symbols still reflect the state before the transformation.
In other words: prettyprint is lying atm.